requirements.txt


# Flask==3.0.3
# ultralytics==8.3.0
# torch==2.1.0
# torchvision==0.16.0
# Pillow==10.0.1
# gunicorn==21.2.0
# numpy==1.26.4
# opencv-python-headless
# --extra-index-url https://download.pytorch.org/whl/cpu


# Flask==3.0.3
# ultralytics==8.1.0
# torch==2.1.0
# torchvision==0.16.0
# Pillow==10.2.0
# gunicorn==21.2.0
# numpy==1.26.4

# Flask==3.0.3
# ultralytics==8.0.196
# torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
# torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cpu
# Pillow==10.0.1
# gunicorn==21.2.0
# numpy==1.26.4
# opencv-python-headless==4.8.0.74
# onnxruntime==1.18.0

Flask==3.0.3
ultralytics==8.3.0
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cpu
Pillow==10.0.1
gunicorn==21.2.0
numpy==1.26.4
opencv-python-headless==4.8.0.74
fastapi==0.109.0
uvicorn==0.29.0
onnxruntime==1.18.0
python-multipart==0.0.6   ← ADD THIS





